+++ spec/formulaic/label_spec.rb
@@ -9,12 +9,70 @@ describe Formulaic::Label do
     expect(label(:user, :first_name)).to eq "First name"
   end
 
-  it "uses a translation if available" do
-    I18n.backend.store_translations(:en, { simple_form: { labels: { user: { name: "Translated" } } } } )
+  context "when simple_form translations are available" do
+    it "uses simple_form translation" do
+      I18n.reload!
+      I18n.backend.store_translations(:en, {
+        simple_form: {
+          labels: {
+            user: {
+              name: "Translated"
+            },
+          },
+        },
+      })
 
-    expect(label(:user, :name)).to eq("Translated")
+      expect(label(:user, :name)).to eq("Translated")
+    end
+  end
 
-    I18n.backend.store_translations(:en, { simple_form: { labels: { user: { name: nil } } } } )
+  context "when helpers.label.<model> translations are available" do
+    it "uses helpers.label translation" do
+      I18n.reload!
+      I18n.backend.store_translations(:en, {
+        helpers: {
+          label: {
+            user: {
+              name: "Translated"
+            },
+          },
+        },
+      })
+
+      expect(label(:user, :name)).to eq("Translated")
+    end
+  end
+
+  context "when helpers.label.text default translations are available" do
+    it "uses helpers.label.text translations" do
+      I18n.reload!
+      I18n.backend.store_translations(:en, {
+        helpers: {
+          label: {
+            text: "Default",
+          },
+        },
+      })
+
+      expect(label(:user, :name)).to eq("Default")
+    end
+  end
+
+  context "when activerecord.attributes.<model> translations are available" do
+    it "uses activerecord.attributes translations" do
+      I18n.reload!
+      I18n.backend.store_translations(:en, {
+        activerecord: {
+          attributes: {
+            user: {
+              name: "Translated"
+            },
+          },
+        },
+      })
+
+      expect(label(:user, :name)).to eq("Translated")
+    end
   end
 
   it "should leave cases alone" do